Service: track kernel-registered addresses, fix duplicates
If a destination gets restarted, the OnAvahiNewObject() call may occur
before the OnAvahiRemoveObject(), leading to attempts to register a
new address again, which will fail with "EEXIST".  On the following
OnAvahiRemoveObject() call, the destination will be removed from the
kernel, and is now gone forever (until either KLB or the destination
is restarted).

A new std::map now tracks which addresses are registered, and if
OnAvahiNewObject() gets called with an address that is already
registered, KLB removes the old one first and then adds the new one
(which should be identical most of the time).  The following
OnAvahiRemoveObject() call is a no-op, because the "destinations" item
does not exist anymore at this point.